proof of concept network multiplayer
Server accepts two connections and creates two NetworkPlayers then
starts a GameManager with those players.

NetworkPlayer gets a move by sending a message to the Client and reading
its reply. It will then tell the server to send the move to the other
player.

Client keeps track of its own Board and a Player. Waits for messages
from the server and replies with a move from its Player or by playing
the opponent's move on its board.