update protocol to handle client disconnecting
Server checks if the player is still connected before adding it to the
player pool.

NetworkGameManager only allows one illegal move since the client should
verify moves. No longer uses a separate thread to run the GameManager.
If the game ends successfully, the NetworkGameManager will send a
REMATCH message to the players. Otherwise, it will send a PING message
to the players and close the connection if there is any problem. If a
player disconnects, the opponent will be notified with a DISCONNECTED
message.

NetworkPlayer checks if the reply from the client is null.

Client handles the new protocol by checking for a PING or DISCONNECTED
message during a game. run method will now continuously loop unless
there is an IOException (instead of only looping when the client
confirms a rematch).