Minor changes to core GUI.
Use Canvas instead of GridPane for BoardPanel.

Fix bug where an exception is thrown when the win condition slider is moved
past the maximum valid value. The bug was caused because when reseting the
slider to the maximum valid value, it was previously done in the JavaFX
Application thread at some unknown future time. This means that it was
possible for the value to change and the board to be reset before the slider
was reset, leading to an illegal argument since the win condition would be
greater than the maximum. The change now occurs in the same thread that resets
the board, so it should no longer be possible to have an invalid win condition slider value before the board is reset.